+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
-## Next
+## Next (1.2.0)
 
-*
+Breaking changes:
+* [#22](https://github.com/rnett/krosstalk/pull/22) Update Ktor auth scopes.  Changes Basic to use Ktor's credential class and adds Bearer scopes.
+  Only breaks if you used `com.rnett.krosstalk.ktor.client.auth.BasicCredentials` directly.
 
 ## 1.1.2